src/backend/requirements.txt
attr==0.3.1
attrs==21.4.0
Django==4.0.3
djangorestframework==3.13.1
